STOP CRASHING DAMMIT!
Added a check for null rock in BlockRockVariant#get (see #39). This should never be null, it is a bug, but it is a bug deeply buried in worldgen code. Until further notice I am going to do this, which should stop it from occuring . (Although granite may appear sporadically in the second layer.)