Add barrel recipe for salt water. Closes TerraFirmaCraft/TerraFirmaCraft#643
This barrel recipe creates salt water at a ratio of 20 salt powder and a full barrel of fresh water to a barrel of salt water. This allows you to brine about 80 pieces of food with the same amount of salt powder needed to salt 20 pieces of food. That's much more brine than in reality, but it's more practical given the low amount of salt powder yielded by mining/grinding rock salt.