Resolves #1347
- Also adds faster rate of cool down when its a cooler liquid than water
- If the item is on top of a block that has ice, packed ice or snow material, it also cools down accordingly


This implementation utilizes EntityJoinWorldEvent and ItemExpireEvent, they go hand-in-hand with each other. 

By manually editing the tag of the EntityItem upon spawning, we can identify what EntityItems should be subjected to the cooling process. We also apply an extremely short lifespan to mimic the speed of the barrel recipe, albeit slightly longer (half a second, but modifiable via config). Then all the checks are done in ItemExpireEvent to set a new cooler temperature depending on if the conditions are met. 

First of all, if temperature is 0 or less, then nothing needs to be done and the original lifespan is restored/added on.

If no conditions are met, the same, short length of lifespan is added on so a quick check can be done once again in case the condition will be met later. Now if, this has been repeated for a long time up until it meets the item's normal lifespan, it will despawn.

Finally, if a condition is met, we extend its lifespan once again so more checks can be done to cool down even further.