Make world gen 8.7% faster.
Revert carvers running during features chunk status. It was causing too many problems as features bleeding into nearby chunks were being undercut with caves once generated. As a result, restrict "createWaterAdjacencyMask" to using only the chunk in question, reducing the amount of getBlockState calls by ~35%, and allowing it to directly access the ChunkPrimer and ChunkSection for more performance improvements. Carvers are showing minor issues with water intersections along chunk borders - this is an unavoidable consequence and is the lesser of two evils.