Fix issues raised in the review: - String ore changed to ItemStack vein; - Encapsulated event attributes; - Added Side attribute to event; - Added more performant enumeration lookup; - Serialize BlockPos as long instead;
Also moved the classification of the ProspectResult ( score -> type ) from a bunch of loose logic inside a propick method to a proper method in ProspectResult that classifies itself. If this class ends up being used elsewhere, for whatever reason, we have the guarantee that the classification will be consistent accross the board.