Add abstraction layers for various capability interactions. This is a future-proofing change, that adds new API ways to interact with TFC provided capabilities, on various objects (primarily `ItemStack`s).
Goals of this refactor are to reduce the exposure of `getCapability()`, `Capability<?>`, and `LazyOptional`. As such, a lot of usage patterns were migrated to simpler helper methods, or a nullable `get()` method. Documentation was added and rewritten for various places in the capability API.

This process was done for `IHeat`, `IFood`, `Forging`, `WorldTracker`, `PlayerData`, `IItemSize`, and `ChunkData`.

Closes #2634