Bloomery rework + Test framework
- Input and catalyst (charcoal) stacks are merged, and can store up to 16/layer of mixed
- Output is based on a ratio, so 100 mB Iron + 2 Charcoal = 1 Bloom - slight buff for using low grade ore, and the extra charcoal efficiency from denser iron is no more
- Updated guide book, and added bloomery game tests

Added a new game test framework using `GameTestAssertions`, which replaces usages of `GameTestHelper`. The latter, being mojang, is not extensible, and we cannot easily add our own helper methods (of which I would like many). So, a new assertion model, with some nice features, and the ability to extend it at will. Behind the scenes it is abusing `ThreadLocal<GameTestHelper>` to remove the need to handle context around.

Bloomery rework. Input and catalyst stacks are combined in the same capacity, and the recipe is based on the ratio of input fluid to catalyst. Currently set at 100mB + 2 Charcoal = 1 Bloom, which is a slight buff for low quality ore, and removes the fuel efficiency gain from casting ingots which was ultimately too overpowered vs. the effort and fuel it required to do.

Ported all existing tests to use the new framework.