Remove antique atlas support (Closes #2664)
Antique Atlas itself is not updated past 1.18.2, however it is maintained in a fork for 1.20.1, which has multiple "rewrite" versions. The fork is written as a fabric mod, and supports Forge through connector + FFAPI. This introduces several problems for supporting it.

It is next-to-impossible to get a working dependency of Antique Atlas into a dev env - it does not remap correctly (being either in Yarn or Intermediary), and when loaded, sources are somewhat broken. Connector also does not seem to want to launch from a dev env, meaning the only environment I can set up is a separate instance with a built TFC and resource pack. This cannot be easily connected to a remote debugger with the correct sources available.

The Antique Atlas fork is not documented in what format or resources it changes - nor was I able to discern by trial and error, how to supply our own biomes. The presence of these atlas files seem to have no effect whatsoever, nor did anything I tried (using both the original Antique Atlas documentation, and digging through the fork's source code on various branches, as a reference. Even when it _seemed_ I was loading something (?) successfully, the map was still visually, completely broken.

If, at some point in the future, we can depend on this as a Forge mod (which basically means the original project needs to be maintained, or forked properly - not as a Fabric-only port), and there is a means to provide support, this can be revisited at that time. For now, you served us well (probably), so long Antique Atlas.